+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 ## [v4.1.2] - 2023.04.14
 ### Fix
+- Export preview
 - Export
   - Fix problem with attribute values which are null in feeds with attributes that contain a number in the code
 ### Change